requirements.txt
--allow-all-external
--allow-unverified PIL
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
boto==2.32.1
celery==3.1.19
cffi==0.8.6
chardet==2.1.1
cryptography==0.6
cssmin==0.2.0
cssselect==0.9.1
Django==1.5.12
django-celery==3.1.17
django-devserver==0.8.0
django-extensions==1.3.11
django-mailgun==0.8.0
django-memcached-pool==0.5
django-piston==0.2.3
django-storages==1.1.8
docutils==0.12
eventlet==0.15.2
factory-boy==2.6.0
fake-factory==0.5.3
fpconst==0.7.2
funcsigs==0.4
functools32==3.2.3.post2
futures==3.0.3
google-api-python-client==1.2
greenlet==0.4.3
gunicorn==0.13.4
httmock==1.2.4
httplib2==0.9.2
johnny-cache==1.6.1a0
jsmin==2.0.11
jws==0.1.2
kombu==3.0.30
librabbitmq==1.6.1
lxml==3.5.0
medusa==0.5.4
meld3==0.6.7
mock==1.3.0
MySQL-python==1.2.5
pbr==1.8.1
PIL==1.1.7
pycparser==2.10
pyOpenSSL==0.14
python-memcached==1.53
pytz==2015.7
requests==2.9.1
shortuuid==0.4.3
six==1.8.0
SOAPpy==0.12.1
South==1.0
supervisor==3.0a8
umemcache==1.6.3
wheel==0.24.0
wstools==0.4.3
